Safe migration/seeder implemented

+++ database/migrations/2019_03_31_073444_create_role_users_table.php
@@ -13,13 +13,15 @@ class CreateRoleUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('role_users', function (Blueprint $table) {
-            $table->uuid('user_id');
-            $table->unsignedInteger('role_id');
+        if(!Schema::hasTable('role_users')) {
+            Schema::create('role_users', function (Blueprint $table) {
+                $table->uuid('user_id');
+                $table->unsignedInteger('role_id');
 
-            $table->foreign('user_id')->references('id')->on('users')->onUpdate('CASCADE')->onDelete('CASCADE');
-            $table->foreign('role_id')->references('id')->on('roles')->onUpdate('CASCADE')->onDelete('CASCADE');
-        });
+                $table->foreign('user_id')->references('id')->on('users')->onUpdate('CASCADE')->onDelete('CASCADE');
+                $table->foreign('role_id')->references('id')->on('roles')->onUpdate('CASCADE')->onDelete('CASCADE');
+            });
+        }
     }
 
     /**
