+++ database/migrations/2019_04_18_051908_create_sys_defaults_table.php
@@ -13,20 +13,22 @@ class CreateSysDefaultsTable extends Migration
      */
     public function up()
     {
-        Schema::create('sys_defaults', function (Blueprint $table) {
-            $table->uuid('id')->primary();
+        if(!Schema::hasTable('sys_defaults')) {
+            Schema::create('sys_defaults', function (Blueprint $table) {
+                $table->uuid('id')->primary();
 
-            // default event
-            $table->uuid('event_id')->nullable();
+                // default event
+                $table->uuid('event_id')->nullable();
 
-            // default announcement
-            $table->uuid('announcement_id')->nullable();
+                // default announcement
+                $table->uuid('announcement_id')->nullable();
 
-            $table->timestamps();
+                $table->timestamps();
 
-            $table->foreign('event_id')->references('id')->on('events')->onDelete('CASCADE');
-            $table->foreign('announcement_id')->references('id')->on('announcements')->onDelete('CASCADE');
-        });
+                $table->foreign('event_id')->references('id')->on('events')->onDelete('CASCADE');
+                $table->foreign('announcement_id')->references('id')->on('announcements')->onDelete('CASCADE');
+            });
+        }
     }
 
     /**
