+++ database/migrations/2019_06_09_181043_create_transactions_table.php
@@ -13,29 +13,31 @@ class CreateTransactionsTable extends Migration
      */
     public function up()
     {
-        Schema::create('transactions', function (Blueprint $table) {
-            $table->uuid('id')->primary();
+        if(!Schema::hasTable('transactions')) {
+            Schema::create('transactions', function (Blueprint $table) {
+                $table->uuid('id')->primary();
 
-            $table->dateTime('date')->nullable();
-            $table->uuid('customer_id')->nullable();
-            $table->uuid('user_id')->nullable()->comment('user who saved the transaction');
-            $table->string('staff_name')->nullable()->comment('static name of user who saved the transaction');
-            $table->string('job_order')->nullable();
-            $table->boolean('saved')->default(false);
-            $table->string('or_number')->nullable();
+                $table->dateTime('date')->nullable();
+                $table->uuid('customer_id')->nullable();
+                $table->uuid('user_id')->nullable()->comment('user who saved the transaction');
+                $table->string('staff_name')->nullable()->comment('static name of user who saved the transaction');
+                $table->string('job_order')->nullable();
+                $table->boolean('saved')->default(false);
+                $table->string('or_number')->nullable();
 
-            $table->string('customer_name');
-            $table->float('total_price')->default(0);
-            $table->dateTime('date_paid')->nullable();
+                $table->string('customer_name');
+                $table->float('total_price')->default(0);
+                $table->dateTime('date_paid')->nullable();
 
-            $table->softDeletes();
-            $table->timestamps();
-            $table->timestamp('synched')->nullable();
+                $table->softDeletes();
+                $table->timestamps();
+                $table->timestamp('synched')->nullable();
 
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('CASCADE')->onUpdate('CASCADE');
-            $table->foreign('customer_id')->references('id')->on('customers')->onDelete('CASCADE');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('CASCADE')->onUpdate('CASCADE');
+                $table->foreign('customer_id')->references('id')->on('customers')->onDelete('CASCADE');
 
-        });
+            });
+        }
     }
 
     /**
