+++ database/migrations/2019_07_13_180812_create_rfid_cards_table.php
@@ -13,22 +13,24 @@ class CreateRfidCardsTable extends Migration
      */
     public function up()
     {
-        Schema::create('rfid_cards', function (Blueprint $table) {
-            $table->uuid('id')->primary();
+        if(!Schema::hasTable('rfid_cards')) {
+            Schema::create('rfid_cards', function (Blueprint $table) {
+                $table->uuid('id')->primary();
 
-            $table->string('rfid')->unique()->nullable();
-            $table->double('balance')->default(0);
-            $table->uuid('customer_id')->nullable()->comment('If issued to customer');
-            $table->uuid('user_id')->nullable()->comment('For master card');
-            $table->char('card_type')->default('c')->comment('c = customer, u = user');
+                $table->string('rfid')->unique()->nullable();
+                $table->double('balance')->default(0);
+                $table->uuid('customer_id')->nullable()->comment('If issued to customer');
+                $table->uuid('user_id')->nullable()->comment('For master card');
+                $table->char('card_type')->default('c')->comment('c = customer, u = user');
 
-            $table->timestamps();
-            $table->softDeletes();
-            $table->timestamp('synched')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
+                $table->timestamp('synched')->nullable();
 
-            $table->foreign('customer_id')->references('id')->on('customers')->onDelete('CASCADE');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('CASCADE')->onUpdate('CASCADE');
-        });
+                $table->foreign('customer_id')->references('id')->on('customers')->onDelete('CASCADE');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('CASCADE')->onUpdate('CASCADE');
+            });
+        }
     }
 
     /**
