+++ database/migrations/2019_07_16_080235_create_rfid_load_transactions_table.php
@@ -13,26 +13,28 @@ class CreateRfidLoadTransactionsTable extends Migration
      */
     public function up()
     {
-        Schema::create('rfid_load_transactions', function (Blueprint $table) {
-            $table->uuid('id')->primary();
+        if(!Schema::hasTable('rfid_load_transactions')) {
+            Schema::create('rfid_load_transactions', function (Blueprint $table) {
+                $table->uuid('id')->primary();
 
-            $table->uuid('rfid_card_id')->nullable();
-            $table->string('customer_name');
-            $table->string('rfid');
-            $table->double('amount')->nullable();
-            $table->double('remaining_balance')->nullable()->comment('Remaining balance before loading.');
-            $table->double('current_balance')->nullable()->comment('Balance after loading.');
-            $table->double('cash')->nullable()->comment('Amount paid.');
-            $table->double('change')->nullable();
-            $table->string('staff_name')->nullable();
-            $table->text('remarks')->nullable();
+                $table->uuid('rfid_card_id')->nullable();
+                $table->string('customer_name');
+                $table->string('rfid');
+                $table->double('amount')->nullable();
+                $table->double('remaining_balance')->nullable()->comment('Remaining balance before loading.');
+                $table->double('current_balance')->nullable()->comment('Balance after loading.');
+                $table->double('cash')->nullable()->comment('Amount paid.');
+                $table->double('change')->nullable();
+                $table->string('staff_name')->nullable();
+                $table->text('remarks')->nullable();
 
-            $table->timestamps();
-            $table->softDeletes();
-            $table->timestamp('synched')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
+                $table->timestamp('synched')->nullable();
 
-            $table->foreign('rfid_card_id')->references('id')->on('rfid_cards')->onDelete('CASCADE')->onUpdate('CASCADE');
-        });
+                $table->foreign('rfid_card_id')->references('id')->on('rfid_cards')->onDelete('CASCADE')->onUpdate('CASCADE');
+            });
+        }
     }
 
     /**
