+++ database/migrations/2019_10_22_234331_create_clients_table.php
@@ -13,16 +13,18 @@ class CreateClientsTable extends Migration
      */
     public function up()
     {
-        Schema::create('clients', function (Blueprint $table) {
-            $table->uuid('user_id')->nullable();
-            $table->string('shop_name')->nullable();
-            $table->string('shop_email')->nullable();
-            $table->string('shop_number')->nullable();
-            $table->string('address')->nullable();
-            $table->timestamps();
+        if(!Schema::hasTable('clients')) {
+            Schema::create('clients', function (Blueprint $table) {
+                $table->uuid('user_id')->nullable();
+                $table->string('shop_name')->nullable();
+                $table->string('shop_email')->nullable();
+                $table->string('shop_number')->nullable();
+                $table->string('address')->nullable();
+                $table->timestamps();
 
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('CASCADE')->onUpdate('CASCADE');
-        });
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('CASCADE')->onUpdate('CASCADE');
+            });
+        }
     }
 
     /**
