+++ database/migrations/2020_01_02_165210_create_service_transaction_items_table.php
@@ -13,32 +13,34 @@ class CreateServiceTransactionItemsTable extends Migration
      */
     public function up()
     {
-        Schema::create('service_transaction_items', function (Blueprint $table) {
-            $table->uuid('id')->primary();
+        if(!Schema::hasTable('service_transaction_items')) {
+            Schema::create('service_transaction_items', function (Blueprint $table) {
+                $table->uuid('id')->primary();
 
-            $table->uuid('transaction_id');
+                $table->uuid('transaction_id');
 
-            $table->string('name');
-            $table->double('price')->default(0);
-            $table->string('category')->comment('washing, drying, other, full');
-            $table->float('earning_points')->default(0);
+                $table->string('name');
+                $table->double('price')->default(0);
+                $table->string('category')->comment('washing, drying, other, full');
+                $table->float('earning_points')->default(0);
 
-            $table->uuid('washing_service_id')->nullable();
-            $table->uuid('drying_service_id')->nullable();
-            $table->uuid('other_service_id')->nullable();
-            $table->uuid('full_service_id')->nullable();
-            $table->boolean('saved')->default(false);
+                $table->uuid('washing_service_id')->nullable();
+                $table->uuid('drying_service_id')->nullable();
+                $table->uuid('other_service_id')->nullable();
+                $table->uuid('full_service_id')->nullable();
+                $table->boolean('saved')->default(false);
 
-            $table->timestamps();
-            $table->softDeletes();
-            $table->timestamp('synched')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
+                $table->timestamp('synched')->nullable();
 
-            $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('CASCADE')->onUpdate('CASCADE');
-            $table->foreign('washing_service_id')->references('id')->on('washing_services')->onDelete('CASCADE')->onUpdate('CASCADE');
-            $table->foreign('drying_service_id')->references('id')->on('drying_services')->onDelete('CASCADE')->onUpdate('CASCADE');
-            $table->foreign('other_service_id')->references('id')->on('other_services')->onDelete('CASCADE')->onUpdate('CASCADE');
-            $table->foreign('full_service_id')->references('id')->on('full_services')->onDelete('CASCADE')->onUpdate('CASCADE');
-        });
+                $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('CASCADE')->onUpdate('CASCADE');
+                $table->foreign('washing_service_id')->references('id')->on('washing_services')->onDelete('CASCADE')->onUpdate('CASCADE');
+                $table->foreign('drying_service_id')->references('id')->on('drying_services')->onDelete('CASCADE')->onUpdate('CASCADE');
+                $table->foreign('other_service_id')->references('id')->on('other_services')->onDelete('CASCADE')->onUpdate('CASCADE');
+                $table->foreign('full_service_id')->references('id')->on('full_services')->onDelete('CASCADE')->onUpdate('CASCADE');
+            });
+        }
     }
 
     /**
