+++ database/migrations/2020_01_05_165654_create_customer_washes_table.php
@@ -13,28 +13,30 @@ class CreateCustomerWashesTable extends Migration
      */
     public function up()
     {
-        Schema::create('customer_washes', function (Blueprint $table) {
-            $table->uuid('id')->primary();
+        if(!Schema::hasTable('customer_washes')) {
+            Schema::create('customer_washes', function (Blueprint $table) {
+                $table->uuid('id')->primary();
 
-            $table->uuid('customer_id');
-            $table->uuid('service_transaction_item_id');
-            $table->string('service_name');
-            $table->string('washer_name')->nullable();
-            $table->string('machine_type')->comment('TITAN, REGULAR');
-            $table->integer('pulse_count');
-            $table->integer('minutes');
-            $table->float('price')->default(0);
-            $table->dateTime('used')->nullable();
-            $table->string('staff_name')->nullable()->comment('staff who activates the service');
-            $table->string('job_order');
+                $table->uuid('customer_id');
+                $table->uuid('service_transaction_item_id');
+                $table->string('service_name');
+                $table->string('washer_name')->nullable();
+                $table->string('machine_type')->comment('TITAN, REGULAR');
+                $table->integer('pulse_count');
+                $table->integer('minutes');
+                $table->float('price')->default(0);
+                $table->dateTime('used')->nullable();
+                $table->string('staff_name')->nullable()->comment('staff who activates the service');
+                $table->string('job_order');
 
-            $table->timestamps();
-            $table->softDeletes();
-            $table->timestamp('synched')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
+                $table->timestamp('synched')->nullable();
 
-            $table->foreign('customer_id')->references('id')->on('customers')->onDelete('CASCADE')->onUpdate('CASCADE');
-            $table->foreign('service_transaction_item_id')->references('id')->on('service_transaction_items')->onDelete('CASCADE')->onUpdate('CASCADE');
-        });
+                $table->foreign('customer_id')->references('id')->on('customers')->onDelete('CASCADE')->onUpdate('CASCADE');
+                $table->foreign('service_transaction_item_id')->references('id')->on('service_transaction_items')->onDelete('CASCADE')->onUpdate('CASCADE');
+            });
+        }
     }
 
     /**
