+++ database/migrations/2020_01_13_173031_create_machine_usages_table.php
@@ -13,20 +13,22 @@ class CreateMachineUsagesTable extends Migration
      */
     public function up()
     {
-        Schema::create('machine_usages', function (Blueprint $table) {
-            $table->uuid('id')->primary();
+        if(!Schema::hasTable('machine_usages')) {
+            Schema::create('machine_usages', function (Blueprint $table) {
+                $table->uuid('id')->primary();
 
-            $table->uuid('machine_id');
-            $table->string('customer_name')->nullable();
-            $table->double('minutes');
-            $table->string('activation_type')->conmment('card, remote')->nullable();
-            $table->float('price')->default(0);
-            $table->text('remarks')->nullable();
-            $table->timestamp('synched')->nullable();
+                $table->uuid('machine_id');
+                $table->string('customer_name')->nullable();
+                $table->double('minutes');
+                $table->string('activation_type')->conmment('card, remote')->nullable();
+                $table->float('price')->default(0);
+                $table->text('remarks')->nullable();
+                $table->timestamp('synched')->nullable();
 
-            $table->timestamps();
-            $table->foreign('machine_id')->references('id')->on('machines')->onDelete('CASCADE')->onUpdate('CASCADE');
-        });
+                $table->timestamps();
+                $table->foreign('machine_id')->references('id')->on('machines')->onDelete('CASCADE')->onUpdate('CASCADE');
+            });
+        }
     }
 
     /**
